Adjustments to Games and Adding Talk
Made adjustments to the games class such as on load of the page pulling in the pet's statistics instead of on call. Also finally implemented the game.talk() method which will take what the output of the game is and phrase it like the pet would say. It is also capable of responding to the user's input.